change rtlsdr_set_tuner_gain_mode() return code behavoir
rtlsdr_set_tuner_gain_mode() returns 0 on success as before
only for the old modes GAIN_MODE_AGC and GAIN_MODE_MANUAL.
This is needed for backwards compatibility.

For the new modes, return the mode set. If a mode is requested that
is not implemented, use the mode with the highest number instead.